requirements.txt
streamlit
crewai
openai
python-dotenv
requests
